Big integer support for Integer widget (#788)

+++ import_export/widgets.py
@@ -85,7 +85,7 @@ class IntegerWidget(NumberWidget):
     def clean(self, value, row=None, *args, **kwargs):
         if self.is_empty(value):
             return None
-        return int(float(value))
+        return int(Decimal(value))
 
 
 class DecimalWidget(NumberWidget):
